Fix footer positioning
This approach results in a footer that is displayed immediately after
the end of the page content (as opposed to "sticking" to the bottom of
the page even when there is not enough content to fill the viewport). A
"sticky" footer normally requires hacks like extraneous, non-semantic
"wrapper" elements in order to render correctly in all browsers.